Disable not supported error handling in all go versions previous 1.11
