+++ middleware/proxy_1_11_n.go
@@ -1,0 +1,13 @@
+// +build !go1.11
+
+package middleware
+
+import (
+	"github.com/labstack/echo"
+	"net/http"
+	"net/http/httputil"
+)
+
+func proxyHTTP(t *ProxyTarget, c echo.Context, config ProxyConfig) http.Handler {
+	return httputil.NewSingleHostReverseProxy(t.URL)
+}
